+++ partiql-lang/src/main/kotlin/org/partiql/lang/util/impl/ResourceAuthority.kt
@@ -1,4 +1,4 @@
-package org.partiql.lang.partiqlisl
+package org.partiql.lang.util.impl
 
 import com.amazon.ion.IonSystem
 import com.amazon.ion.IonValue
@@ -7,14 +7,20 @@ import com.amazon.ionschema.IonSchemaSystem
 import com.amazon.ionschema.util.CloseableIterator
 import java.io.InputStream
 
-class ResourceAuthority(
-    val rootPackage: String,
-    val classLoader: ClassLoader,
+internal class ResourceAuthority(
+    private val rootPackage: String,
+    private val classLoader: ClassLoader,
     val ion: IonSystem
 ) : Authority {
+
+    companion object {
+        internal fun getResourceAuthority(ion: IonSystem) =
+            ResourceAuthority("org/partiql/schemas", ResourceAuthority::class.java.classLoader, ion)
+    }
+
     override fun iteratorFor(iss: IonSchemaSystem, id: String): CloseableIterator<IonValue> {
         val resourceName = "$rootPackage/$id"
-        var str: InputStream? = classLoader.getResourceAsStream(resourceName)
+        val str: InputStream = classLoader.getResourceAsStream(resourceName)
             ?: error("Failed to load schema with resource name '$resourceName'")
 
         return object : CloseableIterator<IonValue> {
@@ -28,16 +34,11 @@ class ResourceAuthority(
             override fun close() {
                 try {
                     reader?.close()
-                    stream?.close()
+                    stream.close()
                 } finally {
                     reader = null
-                    stream = null
                 }
             }
         }
     }
 }
-
-fun getResourceAuthority(ion: IonSystem) = ResourceAuthority("org/partiql/schemas", ResourceAuthority::class.java.getClassLoader(), ion)
-
-fun loadPartiqlIsl(iss: IonSchemaSystem) = iss.loadSchema("partiql.isl")